requirements.txt
click==8.1.3
Faker==17.3.0
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.2
packaging==23.0
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==1.4.46
typing_extensions==4.5.0
Werkzeug==2.2.3
